+++ src/old/apiURL.ts
@@ -1,4 +1,4 @@
-export const BASE_URL = 'https://dokazovi-be.herokuapp.com/api';
+export const BASE_URL = 'http://localhost:8080/api';
 
 const { origin } = window.location;
 export const OAUTH2_REDIRECT_URI = `${origin}/oauth2/redirect`;
